updated to webpack 4, adjusted all dependencies and config, put all required build files in /dist
