--- index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <head>
-        <script src="./dist/bundle.js"></script>
-    </head>
-    <body>
-        <canvas id="canvas"></canvas>
-    </body>
-    <script>
-        (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-        (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-        m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-        })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-
-        ga('create', 'UA-98365855-2', 'auto');
-        ga('send', 'pageview', location.pathname);
-    </script>
-</html>
