+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -17,13 +18,14 @@
   },
   "homepage": "https://github.com/HarveyD/repeat-after-me#readme",
   "devDependencies": {
-    "css-loader": "^0.28.1",
-    "file-loader": "^0.11.1",
+    "css-loader": "^0.28.11",
+    "file-loader": "^1.1.11",
+    "html-webpack-plugin": "^3.1.0",
     "path": "^0.12.7",
-    "style-loader": "^0.17.0",
-    "ts-loader": "^2.0.3"
-  },
-  "dependencies": {
-    "typescript": "^2.4.2"
+    "style-loader": "^0.20.3",
+    "ts-loader": "^4.1.0",
+    "typescript": "^2.8.1",
+    "webpack": "^4.4.1",
+    "webpack-cli": "^2.0.13"
   }
 }
