+++ tsconfig.json
@@ -1,16 +1,16 @@
 {
-    "compilerOptions": {
-        "outDir": "./dist",
-        "sourceMap": true,
-        "noImplicitAny": true,
-        "module": "commonjs",
-        "target": "es6",
-        "allowJs": true
-    },
-    "exclude": [
-        "node_modules"
-    ],
-    "files": [
-        "./src/app.ts"
-    ]
+	"compilerOptions": {
+		"outDir": "./dist",
+		"sourceMap": true,
+		"noImplicitAny": true,
+		"module": "commonjs",
+		"target": "es6",
+		"allowJs": true
+	},
+	"exclude": [
+		"node_modules"
+	],
+	"files": [
+		"./src/js/app.ts"
+	]
 }
