+++ webpack.config.js
@@ -1,32 +1,44 @@
 const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-    entry: './src/app.ts',
-    output: {
-        path: path.resolve(__dirname, "dist"),
-        filename: "bundle.js",
-        publicPath: './dist/'
-    },
-    resolve: {
-        extensions: ['.ts', '.tsx', '.js']
-    },
-    module: {
-        loaders: [
-            {
-                test: /\.ts?$/,
-                exclude: /node_modules/,
-                loader: "ts-loader"
-            },
-            {
-                test: /\.css$/,
-                exclude: /node_modules/,
-                loader: "style-loader!css-loader"
-            },
-            {
-                test: /\.mp3$/,
-                exclude: /node_modules/,
-                loader: 'file-loader'
-            }
-        ]
-    }
+	entry: './src/js/app.ts',
+	output: {
+		path: path.resolve(__dirname, "dist"),
+		filename: "bundle.js",
+	},
+	plugins: [
+    new HtmlWebpackPlugin({
+			template: './src/index.html'
+		})
+  ],
+	resolve: {
+		extensions: ['.ts', '.tsx', '.js']
+	},
+	module: {
+		rules: [
+			{
+				test: /\.ts?$/,
+				exclude: /node_modules/,
+				use: [
+					{ loader: "ts-loader" }
+				]
+			},
+			{
+				test: /\.css$/,
+				exclude: /node_modules/,
+				use: [
+					"style-loader",
+					"css-loader"
+				]
+			},
+			{
+				test: /\.mp3$/,
+				exclude: /node_modules/,
+				use: [
+					{ loader: 'file-loader' }
+				]
+			}
+		]
+	}
 }
